requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
torch
torchvision
numpy
scikit-learn
pandas
opencv-python
matplotlib
tensorboard
shap
scikit-image
